+++ requirments.txt
@@ -1,2 +1,3 @@
 selenium==3.141.0
 pytest==6.0.1
+flake8==3.8.4
